Move search.py
